Just copy pasted the code from Gemini
Gemini doesn't generate code in one prompt.

Prompts
1.
What’s this Application does?
This is a mathematical based application. This application is design to improve the mental math.
The idea is guess arithmetic problems answers which should some simple math like addiction and subtraction. The user have to guess Answer of the problems

I have some idea for the application. We will provide some trick and technics to calculate the problem with your mind. that way they calculate this problems and make fast response to problems. we suggest the trick and everything after each game section is done.
After each game section is done we will show a graph based result sheet just like the monkeytype.com . that way user can see the result more in a detailed way and also make bit like game like that way we can make it bit addictive too.
To make it more interesting we making a point system just like monkeytype.com for this we need user login and everything for that.

how to build a simplest version of the project?

2.
In this I have to type the answer, We need options method for input provide 4 options to choose and click the correct answer

I was using the Personalization(experimental)) model for this.